CLIENT / organize CreateContentModal folder
